+++ testvectorgen/utils/constants.go
@@ -1,0 +1,41 @@
+package utils
+
+// List of available operators.
+const (
+	NOOP int = iota // No operation, skip query verification in circuit
+	EQ
+	LT
+	GT
+	IN
+	NIN
+)
+
+const TestClaimDocument = `{
+   "@context": [
+     "https://www.w3.org/2018/credentials/v1",
+     "https://w3id.org/citizenship/v1",
+     "https://w3id.org/security/bbs/v1"
+   ],
+   "id": "https://issuer.oidp.uscis.gov/credentials/83627465",
+   "type": ["VerifiableCredential", "PermanentResidentCard"],
+   "issuer": "did:example:489398593",
+   "identifier": 83627465,
+   "name": "Permanent Resident Card",
+   "description": "Government of Example Permanent Resident Card.",
+   "issuanceDate": "2019-12-03T12:19:52Z",
+   "expirationDate": "2029-12-03T12:19:52Z",
+   "credentialSubject": {
+     "id": "did:example:b34ca6cd37bbf23",
+     "type": ["PermanentResident", "Person"],
+     "givenName": "JOHN",
+     "familyName": "SMITH",
+     "gender": "Male",
+     "image": "data:image/png;base64,iVBORw0KGgokJggg==",
+     "residentSince": "2015-01-01",
+     "lprCategory": "C09",
+     "lprNumber": "999-999-999",
+     "commuterClassification": "C1",
+     "birthCountry": "Bahamas",
+     "birthDate": "1958-07-17"
+   }
+ }`
