+++ testvectorgen/utils/utils.go
@@ -1,0 +1,153 @@
+package utils
+
+import (
+	"context"
+	"encoding/hex"
+	"math/big"
+	"os"
+	"strings"
+	"testing"
+	"time"
+
+	core "github.com/iden3/go-iden3-core"
+	"github.com/iden3/go-iden3-crypto/babyjub"
+	"github.com/iden3/go-iden3-crypto/poseidon"
+	"github.com/iden3/go-merkletree-sql/v2"
+	"github.com/iden3/go-schema-processor/merklize"
+)
+
+func PrepareProof(proof *merkletree.Proof) ([]string, NodeAuxValue) {
+	return PrepareSiblingsStr(proof.AllSiblings(), 32), getNodeAuxValue(proof)
+}
+
+func ExtractPubXY(privKHex string) (key *babyjub.PrivateKey, x, y *big.Int) {
+	// Extract pubKey
+	var k babyjub.PrivateKey
+	if _, err := hex.Decode(k[:], []byte(privKHex)); err != nil {
+		panic(err)
+	}
+	pk := k.Public()
+	return &k, pk.X, pk.Y
+}
+
+func DefaultJSONUserClaim(subject core.ID) (*merklize.Merklizer, *core.Claim, error) {
+	mz, err := merklize.MerklizeJSONLD(context.Background(), strings.NewReader(TestClaimDocument))
+	if err != nil {
+		return nil, nil, err
+	}
+
+	var schemaHash core.SchemaHash
+	schemaBytes, err := hex.DecodeString("ce6bb12c96bfd1544c02c289c6b4b987")
+	copy(schemaHash[:], schemaBytes)
+
+	nonce := 10
+
+	claim, err := core.NewClaim(
+		schemaHash,
+		core.WithIndexID(subject),
+		core.WithExpirationDate(time.Unix(1669884010, 0)), //Thu Dec 01 2022 08:40:10 GMT+0000
+		core.WithRevocationNonce(uint64(nonce)),
+		core.WithIndexMerklizedRoot(mz.Root().BigInt()))
+
+	return mz, claim, err
+}
+
+func HashToStr(siblings []*merkletree.Hash) []string {
+	siblingsStr := make([]string, len(siblings))
+	for i, sibling := range siblings {
+		siblingsStr[i] = sibling.BigInt().String()
+	}
+	return siblingsStr
+}
+
+func DefaultUserClaim(subject core.ID) (*core.Claim, error) {
+	dataSlotA, _ := core.NewElemBytesFromInt(big.NewInt(10))
+	nonce := 1
+	var schemaHash core.SchemaHash
+	schemaBytes, err := hex.DecodeString("ce6bb12c96bfd1544c02c289c6b4b987")
+	if err != nil {
+		return nil, err
+	}
+	copy(schemaHash[:], schemaBytes)
+
+	return core.NewClaim(
+		schemaHash,
+		core.WithIndexID(subject),
+		core.WithIndexData(dataSlotA, core.ElemBytes{}),
+		core.WithExpirationDate(time.Unix(1669884010, 0)), //Thu Dec 01 2022 08:40:10 GMT+0000
+		core.WithRevocationNonce(uint64(nonce)))
+
+}
+
+func PrepareStrArray(siblings []string, levels int) []string {
+	// Add the rest of empty levels to the array
+	for i := len(siblings); i < levels; i++ {
+		siblings = append(siblings, "0")
+	}
+	return siblings
+}
+
+func IDFromState(state *big.Int) (*core.ID, error) {
+	typ, err := core.BuildDIDType(core.DIDMethodIden3, core.NoChain, core.NoNetwork)
+	if err != nil {
+		return nil, err
+	}
+	// create new identity
+	return core.IdGenesisFromIdenState(typ, state)
+}
+
+func PrepareSiblingsStr(siblings []*merkletree.Hash, levels int) []string {
+	// siblings := mtproof.AllSiblings()
+	// Add the rest of empty levels to the siblings
+	for i := len(siblings); i < levels; i++ {
+		siblings = append(siblings, &merkletree.HashZero)
+	}
+	return HashToStr(siblings)
+}
+
+func NewAuthClaim(privKHex string) (auth *core.Claim, key *babyjub.PrivateKey, err error) {
+	// extract pubKey
+	key, X, Y := ExtractPubXY(privKHex)
+
+	// create auth claim
+	authClaim, err := AuthClaimFromPubKey(X, Y)
+	if err != nil {
+		return nil, nil, err
+	}
+	return authClaim, key, nil
+}
+
+func AuthClaimFromPubKey(X, Y *big.Int) (*core.Claim, error) {
+	var schemaHash core.SchemaHash
+	//schemaEncodedBytes, _ := hex.DecodeString("ca938857241db9451ea329256b9c06e5") // V1
+	schemaEncodedBytes, _ := hex.DecodeString("013fd3f623559d850fb5b02ff012d0e2") // V2
+	copy(schemaHash[:], schemaEncodedBytes)
+
+	// NOTE: We take nonce as hash of public key to make it random
+	// We don't use random number here because this test vectors will be used for tests
+	// and have randomization inside tests is usually a bad idea
+	revNonce, err := poseidon.Hash([]*big.Int{X})
+	if err != nil {
+		return nil, err
+	}
+
+	return core.NewClaim(schemaHash,
+		core.WithIndexDataInts(X, Y),
+		core.WithRevocationNonce(revNonce.Uint64()))
+}
+
+func SaveTestVector(t *testing.T, fileName string, data string) {
+	t.Helper()
+	path := "testdata/" + fileName + ".json"
+
+	f, err := os.Create(path)
+	defer f.Close()
+	if err != nil {
+		t.Fatalf("Error writing to file %s: %s", path, err)
+	}
+
+	_, err = f.WriteString(data)
+	if err != nil {
+		t.Fatalf("Error writing to file %s: %s", path, err)
+	}
+}
